Fixed some basic syntax stuff
Made some of the code in PowerPlantAgent more concise, (thanks for pointing out the stream().filter() stuff Huan). Reseller agent has some work done to increase its conciseness too.

The base agent changes are not significant in any way, just fixing an issue with when we get json and when we invoke the bookkeeping update. Also getting rid of some stuff that causes flags in intlij for java violations.